See #34. Adds SPI 'test' flag and response.
When an SPI transfer is initiated and the Pi sends the SPI_ECHO_TEST
flag in its payload, the Arduino takes note and loads the register
with the SPI_ECHO_RESPONSE value for the next transmission. The
spi_getbyte() function is then called again, requesting that RESPONSE
value from the Arduino. If this is indeed sent to the Pi, everything
is working as intended.