See #34. Misc data collection adjustments.
Makes variables local to 'data_collection' for the time being because
of issues with pointers passing down through multiple functions. In
this iteration, enough data is stored locally for changes to be
observed and printed to the screen.

Current status:
Pi code reads in changes and can calculate and display pings. The
arduino seems to be passing the data along correctly as well. However,
there are still timing errors, and sometimes the numbers jump to near
2^32 (indicating some kind of overflow, perhaps). Where this is
happening is TBD.