See #34. Fixes bounded buffer issue on Arduino
Fixes issue where the "end" would wrap around to index 0, and then SPI
would also wrap to 0 and start sending whatever was there. Of course,
if data came in during an SPI transfer, this could lead to a race
condition where various parts of two different records could be sent to
the Pi, of course leading to bad data.