See #40. Fixes Pi polling too quickly for SPI "blocks"
Before, the Pi would request the 5 bytes (pinvals + timestamp in 4
steps) back-to-back with no delay in between. This mean that often the
SPI data register was not loaded with the appropriate data in time, and
therefore corrupted data would be sent to the Pi.

Specifically, the pinvals would get loaded into the data register at
some point and would be ready. Then, the Pi would initiate an SPI data
transfer and the pinvals would be sent. As soon as the last byte was
received, the Pi would then start to request the timestamp, which is a
32-bit unsigned long. However, the Arduino was sometimes unable to load
this into the SPI data register in time, as it may have been dealing
with incoming data on the pins. Thus by the time it got back around to
putting the timestamp data into the SPI data register, the Pi would
have already consumed half of that data.

Further, the Pi generally sends only zeros in the 2-way communication
with SPI. These would be shifted into the SPDR, meaning that the
register would be filled with the Pi's zeros when it came around too
quickly the next time. This would lead to sets of zeros in front of
some timestamps, meaning that they were greatly diminished in numerical
value.

With this commit, a short wait() has been added when the Pi makes an
SPI data transfer. It is currently set to 10 microseconds, which should
be enough time for the Arduino to complete a cycle of its main loop()
function and load the next byte into the SPI data register. However, we
may adjust this number a bit as needed and do some further stress tests
to make sure that it is enough of a wait.