See #40. Attempt to use MSB of 'pinvals' to denote stale data
After sending a pinval, the arduino updates that pinval entry so that
the most significant bit (unused until now) is set to a 1. When a
change on the pins is detected and the new pinvals stored, the MSB is
stored as a 0. This way, when the Pi reads a block of data from SPI, it
can tell if it was a block that was sent to it previously (indicated by
a 1 in the MSB of the `pinvals` byte). The Pi can then discard/ignore
this data if needed.

The motivation for this, yet to be documented, is to handle a race
condition in which the "write" end of the buffer catches up with the
"read" end, and new data is stored in that location as an SPI data
transfer is occurring. The specifics of why this works will be detailed
later. However, it should be noted that this still does not entirely
solve the problem at hand.