See #12. Partial work done on refactoring direction tests.
As originally written, the tests for the direction algorithm (in
`test_direction.cpp`) had their own copy of `triangulation_helper()`
defined. At some point, these two functions diverged; that is, the
version in 'direction.cpp' was different from the version in
'test_direction.cpp'. While I have not investigated the full
difference, one striking difference is that they each returned
a different struct. Both structs were called `data`, but they
had different definitions in the two files.

The `data` struct that was used in `direction.cpp` included d, r,
and theta. It has been renamed to `crab_event` to avoid confusion.
The `data` struct that was used in `test-direction.cpp` had
the fields in the other struct as well as x, y, and N. The test
code relied on these additional fields for some calculations.
However, because we want to ensure that we test the "actual"
function, and not copies that may change independently of what
will be run in the main application, it is important that the
two versions of `triangulation_helper()` combined into a single
file that lives in `direction.cpp`. This means reconcilng the
two possible return structs into one that will be used
everywhere in the program. There is some flexibility in how we
go about this.

As of this commit, the duplicate copy of the method in question
has been removed from the test file; unfortunately, this means
that the tests no longer compile, as there are references to
properties of the 'data' struct that no longer exist. We should
have a dialog on how best to move ahead on this issue.